adicionar transacoes, com testes e excecoes adequadas

+++ miniautorizador/src/main/java/desafio/vr/miniautorizador/models/Cartao.java
@@ -23,7 +23,7 @@ public class Cartao {
         return saldo;
     }
 
-    public void setSaldo(float saldo) {
+    public void setSaldo(double saldo) {
         this.saldo = saldo;
     }
